algorithms/medium/621-task-scheduler/src/main/java/ru/saidgadjiev/leetcode/medium/_621/Solution.java
package ru.saidgadjiev.leetcode.medium._621;

/**
 * https://leetcode.com/problems/task-scheduler/
 *
 * ["A","A","A","B","B","B"], n = 2 -> 8
 * A -> B -> idle -> A -> B -> idle -> A -> B
 * There is at least 2 units of time between any two same tasks.
 */
//TODO: Надо разобрать
public class Solution {
    public int leastInterval(char[] tasks, int n) {
        int[] counter = new int[26];
        int max = 0;
        int maxCount = 0;
        for(char task : tasks) {
            counter[task - 'A']++;
            if(max == counter[task - 'A']) {
                maxCount++;
            }
            else if(max < counter[task - 'A']) {
                max = counter[task - 'A'];
                maxCount = 1;
            }
        }

        int partCount = max - 1;
        int partLength = n - (maxCount - 1);
        int emptySlots = partCount * partLength;
        int availableTasks = tasks.length - max * maxCount;
        int idles = Math.max(0, emptySlots - availableTasks);

        return tasks.length + idles;
    }
}
